型を変更

+++ db/schema.rb
@@ -15,7 +15,7 @@ ActiveRecord::Schema.define(version: 2020_02_05_082048) do
   create_table "items", options: "ENGINE=InnoDB DEFAULT CHARSET=utf8", force: :cascade do |t|
     t.string "title", null: false
     t.text "details", null: false
-    t.integer "price", null: false
+    t.bigint "price", null: false
     t.integer "seller_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
